Official Release version 1.4.0

+++ Projects/P-NUCLEO-WB55-NUCLEO/Applications/IKS01A3/MOTENV1/Src/MotionCP_Manager.c
@@ -1,4 +1,4 @@
-
+/* USER CODE BEGIN Header */
 /**
   ******************************************************************************
   * @file    MotionCP_Manager.c
@@ -7,7 +7,7 @@
   ******************************************************************************
   * @attention
   *
-  * Copyright (c) 2023 STMicroelectronics.
+  * Copyright (c) 2024 STMicroelectronics.
   * All rights reserved.
   *
   * This software is licensed under terms that can be found in the LICENSE file
@@ -17,25 +17,27 @@
   ******************************************************************************
   */
 
+/* USER CODE END Header */
+
 /* Includes ------------------------------------------------------------------*/
 #include "MotionCP_Manager.h"
 
 /** @addtogroup MOTION_APPLICATIONS MOTION APPLICATIONS
- * @{
- */
+  * @{
+  */
 
 /** @addtogroup CARRY_POSITION CARRY POSITION
- * @{
- */
+  * @{
+  */
 
 /* Private typedef -----------------------------------------------------------*/
 /* Exported function prototypes ----------------------------------------------*/
 /* Exported functions --------------------------------------------------------*/
 /**
- * @brief  Initialises MotionCP algorithm
- * @param  None
- * @retval None
- */
+  * @brief  Initialises MotionCP algorithm
+  * @param  None
+  * @retval None
+  */
 void MotionCP_manager_init(void)
 {
   char acc_orientation[3];
@@ -50,32 +52,31 @@ void MotionCP_manager_init(void)
 }
 
 /**
- * @brief  Run Carry Position algorithm
- * @param  data_in Structure containing input data
- * @param  data_out Structure containing ouput data
- * @retval None
- */
+  * @brief  Run Carry Position algorithm
+  * @param  data_in Structure containing input data
+  * @param  data_out Structure containing output data
+  * @retval None
+  */
 void MotionCP_manager_run(MCP_input_t *data_in, MCP_output_t *data_out)
 {
   MotionCP_Update(data_in, data_out);
 }
 
 /**
- * @brief  Get the library version
- * @param  version  Library version string (must be array of 35 char)
- * @param  length  Library version string length
- * @retval None
- */
-void MotionCP_manager_get_version(char *version, int *length)
+  * @brief  Get the library version
+  * @param  version  Library version string (must be array of 35 char)
+  * @param  length  Library version string length
+  * @retval None
+  */
+void MotionCP_manager_get_version(char *version, uint8_t *length)
 {
-  *length = (int)MotionCP_GetLibVersion(version);
+  *length = (uint8_t)MotionCP_GetLibVersion(version);
 }
 
 /**
- * @}
- */
+  * @}
+  */
 
 /**
- * @}
- */
-
+  * @}
+  */
