+++ Projects/P-NUCLEO-WB55-NUCLEO/Applications/IKS01A3/MOTENV1/Src/MotionGR_Manager.c
@@ -1,4 +1,4 @@
-
+/* USER CODE BEGIN Header */
 /**
   ******************************************************************************
   * @file    MotionGR_Manager.c
@@ -7,7 +7,7 @@
   ******************************************************************************
   * @attention
   *
-  * Copyright (c) 2023 STMicroelectronics.
+  * Copyright (c) 2024 STMicroelectronics.
   * All rights reserved.
   *
   * This software is licensed under terms that can be found in the LICENSE file
@@ -17,32 +17,33 @@
   ******************************************************************************
   */
 
+/* USER CODE END Header */
+
 /* Includes ------------------------------------------------------------------*/
 #include "MotionGR_Manager.h"
 
 /** @addtogroup MOTION_APPLICATIONS MOTION APPLICATIONS
- * @{
- */
+  * @{
+  */
 
 /** @addtogroup GESTURE_RECOGNITION GESTURE RECOGNITION
- * @{
- */
+  * @{
+  */
 
 /* Private typedef -----------------------------------------------------------*/
 /* Exported function prototypes ----------------------------------------------*/
 /* Exported functions --------------------------------------------------------*/
 /**
- * @brief  Initialize the MotionGR engine
- * @param  None
- * @retval None
- */
+  * @brief  Initialize the MotionGR engine
+  * @param  None
+  * @retval None
+  */
 void MotionGR_manager_init(void)
 {
   char acc_orientation[3];
 
   MotionGR_Initialize();
 
-  /* LSM6DSL */
   acc_orientation[0] = 's';
   acc_orientation[1] = 'e';
   acc_orientation[2] = 'u';
@@ -51,32 +52,31 @@ void MotionGR_manager_init(void)
 }
 
 /**
- * @brief  Run Gesture Recognition algorithm
- * @param  data_in  Structure containing input data
- * @param  data_out Structure containing output data
- * @retval None
- */
+  * @brief  Run Gesture Recognition algorithm
+  * @param  data_in  Structure containing input data
+  * @param  data_out Structure containing output data
+  * @retval None
+  */
 void MotionGR_manager_run(MGR_input_t *data_in, MGR_output_t *data_out)
 {
   MotionGR_Update(data_in, data_out);
 }
 
 /**
- * @brief  Get the library version
- * @param  version  Library version string (must be array of 35 char)
- * @param  length  Library version string length
- * @retval None
- */
-void MotionGR_manager_get_version(char *version, int *length)
+  * @brief  Get the library version
+  * @param  version  Library version string (must be array of 35 char)
+  * @param  length  Library version string length
+  * @retval None
+  */
+void MotionGR_manager_get_version(char *version, uint8_t *length)
 {
-  *length = (int)MotionGR_GetLibVersion(version);
+  *length = (uint8_t)MotionGR_GetLibVersion(version);
 }
 
 /**
- * @}
- */
+  * @}
+  */
 
 /**
- * @}
- */
-
+  * @}
+  */
