+++ Projects/P-NUCLEO-WB55-NUCLEO/Applications/IKS01A3/MOTENV1/Src/MotionPM_Manager.c
@@ -1,4 +1,4 @@
-
+/* USER CODE BEGIN Header */
 /**
   ******************************************************************************
   * @file    MotionPM_Manager.c
@@ -7,7 +7,7 @@
   ******************************************************************************
   * @attention
   *
-  * Copyright (c) 2023 STMicroelectronics.
+  * Copyright (c) 2024 STMicroelectronics.
   * All rights reserved.
   *
   * This software is licensed under terms that can be found in the LICENSE file
@@ -17,57 +17,58 @@
   ******************************************************************************
   */
 
+/* USER CODE END Header */
+
 /* Includes ------------------------------------------------------------------*/
 #include "MotionPM_Manager.h"
 
 /** @addtogroup MOTION_APPLICATIONS MOTION APPLICATIONS
- * @{
- */
+  * @{
+  */
 
 /** @addtogroup PEDOMETER PEDOMETER
- * @{
- */
+  * @{
+  */
 
 /* Private typedef -----------------------------------------------------------*/
 /* Exported function prototypes ----------------------------------------------*/
 /* Exported functions --------------------------------------------------------*/
 /**
- * @brief  Initialize the MotionPM engine
- * @param  None
- * @retval None
- */
+  * @brief  Initialize the MotionPM engine
+  * @param  None
+  * @retval None
+  */
 void MotionPM_manager_init(void)
 {
   MotionPM_Initialize();
 }
 
 /**
- * @brief  Run Pedometer algorithm
- * @param  data_in  Structure containing input data
- * @param  data_out Structure containing output data
- * @retval None
- */
+  * @brief  Run Pedometer algorithm
+  * @param  data_in  Structure containing input data
+  * @param  data_out Structure containing output data
+  * @retval None
+  */
 void MotionPM_manager_run(MPM_input_t *data_in, MPM_output_t *data_out)
 {
   MotionPM_Update(data_in, data_out);
 }
 
 /**
- * @brief  Get the library version
- * @param  version  Library version string (must be array of 35 char)
- * @param  length  Library version string length
- * @retval None
- */
-void MotionPM_manager_get_version(char *version, int *length)
+  * @brief  Get the library version
+  * @param  version  Library version string (must be array of 35 char)
+  * @param  length  Library version string length
+  * @retval None
+  */
+void MotionPM_manager_get_version(char *version, uint8_t *length)
 {
-  *length = (int)MotionPM_GetLibVersion(version);
+  *length = (uint8_t)MotionPM_GetLibVersion(version);
 }
 
 /**
- * @}
- */
+  * @}
+  */
 
 /**
- * @}
- */
-
+  * @}
+  */
