+++ Projects/P-NUCLEO-WB55-NUCLEO/Applications/IKS01A3/MOTENV1_OTA/Inc/MotionID_Manager.h
@@ -1,4 +1,4 @@
-
+/* USER CODE BEGIN Header */
 /**
   ******************************************************************************
   * @file    MotionID_Manager.h
@@ -7,7 +7,7 @@
   ******************************************************************************
   * @attention
   *
-  * Copyright (c) 2023 STMicroelectronics.
+  * Copyright (c) 2024 STMicroelectronics.
   * All rights reserved.
   *
   * This software is licensed under terms that can be found in the LICENSE file
@@ -16,6 +16,8 @@
   *
   ******************************************************************************
   */
+
+/* USER CODE END Header */
 
 /* Define to prevent recursive inclusion -------------------------------------*/
 #ifndef MOTIONID_MANAGER_H
@@ -37,7 +39,7 @@ extern "C" {
 /* Exported Functions Prototypes ---------------------------------------------*/
 void MotionID_manager_init(void);
 void MotionID_manager_run(MID_input_t *data_in, MID_output_t *data_out);
-void MotionID_manager_get_version(char *version, int *length);
+void MotionID_manager_get_version(char *version, uint8_t *length);
 
 #ifdef __cplusplus
 }
