+++ Projects/P-NUCLEO-WB55-NUCLEO/Applications/IKS4A1/MOTENV1_OTA/Src/MotionAR_Manager.c
@@ -1,4 +1,4 @@
-
+/* USER CODE BEGIN Header */
 /**
   ******************************************************************************
   * @file    MotionAR_Manager.c
@@ -7,7 +7,7 @@
   ******************************************************************************
   * @attention
   *
-  * Copyright (c) 2023 STMicroelectronics.
+  * Copyright (c) 2024 STMicroelectronics.
   * All rights reserved.
   *
   * This software is licensed under terms that can be found in the LICENSE file
@@ -17,24 +17,26 @@
   ******************************************************************************
   */
 
+/* USER CODE END Header */
+
 /* Includes ------------------------------------------------------------------*/
 #include <limits.h>
 #include "MotionAR_Manager.h"
 
 /** @addtogroup MOTION_APPLICATIONS MOTION APPLICATIONS
- * @{
- */
+  * @{
+  */
 
 /** @addtogroup ACTIVITY_RECOGNITION ACTIVITY RECOGNITION
- * @{
- */
+  * @{
+  */
 
 /* Exported functions --------------------------------------------------------*/
 /**
- * @brief  Initialize the MotionAR engine
- * @param  None
- * @retval None
- */
+  * @brief  Initialize the MotionAR engine
+  * @param  None
+  * @retval None
+  */
 void MotionAR_manager_init(void)
 {
   char acc_orientation[3];
@@ -49,34 +51,33 @@ void MotionAR_manager_init(void)
 }
 
 /**
- * @brief  Run Activity Recognition algorithm. This function collects and scale
- *         data from accelerometer and calls the Activity Recognition Algo
- * @param  data_in  Structure containing input data
- * @param  data_out Structure containing output data
- * @param  timestamp  Timestamp in [ms]
- * @retval None
- */
+  * @brief  Run Activity Recognition algorithm. This function collects and scale
+  *         data from accelerometer and calls the Activity Recognition Algo
+  * @param  data_in  Structure containing input data
+  * @param  data_out Structure containing output data
+  * @param  timestamp  Timestamp in [ms]
+  * @retval None
+  */
 void MotionAR_manager_run(MAR_input_t *data_in, MAR_output_t *data_out, uint32_t timestamp)
 {
   MotionAR_Update(data_in, data_out, (long int)(timestamp & LONG_MAX));
 }
 
 /**
- * @brief  Get the library version
- * @param  version  Library version string (must be array of 35 char)
- * @param  length  Library version string length
- * @retval None
- */
-void MotionAR_manager_get_version(char *version, int *length)
+  * @brief  Get the library version
+  * @param  version  Library version string (must be array of 35 char)
+  * @param  length  Library version string length
+  * @retval None
+  */
+void MotionAR_manager_get_version(char *version, uint8_t *length)
 {
-  *length = (int)MotionAR_GetLibVersion(version);
+  *length = (uint8_t)MotionAR_GetLibVersion(version);
 }
 
 /**
- * @}
- */
+  * @}
+  */
 
 /**
- * @}
- */
-
+  * @}
+  */
